+++ order-service/src/main/java/com/example/orderservice/domain/exception/NotExistsOrderException.java
@@ -1,0 +1,11 @@
+package com.example.orderservice.domain.exception;
+
+import com.example.orderservice.global.error.ErrorCode;
+import com.example.orderservice.global.exception.BaseException;
+
+public class NotExistsOrderException extends BaseException {
+
+    public NotExistsOrderException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+}
