+++ order-service/src/main/java/com/example/orderservice/domain/model/response/CreateOrderResponse.java
@@ -1,19 +1,19 @@
-package com.example.orderservice.model.response;
+package com.example.orderservice.domain.model.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import jakarta.persistence.criteria.CriteriaBuilder.In;
+import com.fasterxml.jackson.annotation.JsonInclude.Include;
 import java.util.Date;
 import lombok.Data;
 
 @Data
-@JsonInclude(JsonInclude.Include.NON_NULL)
-public class GetOrderResponse {
+@JsonInclude(Include.NON_NULL)
+public class CreateOrderResponse {
 
     private String orderId;
     private String productId;
     private Integer quantity;
     private Integer price;
     private Integer totalPrice;
-    private Date createdAt;
+    private Date createAt;
 
 }
