+++ order-service/src/main/java/com/example/orderservice/domain/model/response/GetOrderResponse.java
@@ -1,19 +1,19 @@
-package com.example.orderservice.model.response;
+package com.example.orderservice.domain.model.response;
 
 import com.fasterxml.jackson.annotation.JsonInclude;
-import com.fasterxml.jackson.annotation.JsonInclude.Include;
+import jakarta.persistence.criteria.CriteriaBuilder.In;
 import java.util.Date;
 import lombok.Data;
 
 @Data
-@JsonInclude(Include.NON_NULL)
-public class CreateOrderResponse {
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class GetOrderResponse {
 
     private String orderId;
     private String productId;
     private Integer quantity;
     private Integer price;
     private Integer totalPrice;
-    private Date createAt;
+    private Date createdAt;
 
 }
