--- order-service/src/main/java/com/example/orderservice/exception/NotExistsOrderException.java
@@ -1,10 +1,0 @@
-package com.example.orderservice.exception;
-
-import com.example.orderservice.error.ErrorCode;
-
-public class NotExistsOrderException extends BaseException {
-
-    public NotExistsOrderException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-}
